Module mmtk::build_info

Structs

MMTK_FULL_BUILD_INFO

Full build info, including MMTk’s name, version, git, and features in the build, such as MMTk 0.14.0 (43e0ce8-dirty, DEFAULT, EXTREME_ASSERTIONS)

MMTK_GIT_VERSION

Git version as short commit hash, such as a96e8f9, or a96e8f9-dirty, or unknown-git-version if MMTk is not built from a git repo.

Constants

MMTK_FEATURES

Comma separated features enabled for this build

MMTK_PKG_VERSION

MMTk crate version such as 0.14.0