Trait mmtk::plan::MutatorContext

pub trait MutatorContext<VM: VMBinding>: Send + 'static {
    fn prepare(&mut self, tls: VMWorkerThread);

    fn release(&mut self, tls: VMWorkerThread);

    fn alloc(
        &mut self, 
        size: usize, 
        align: usize, 
        offset: isize, 
        allocator: AllocationSemantics
    ) -> Address;

    fn post_alloc(
        &mut self, 
        refer: ObjectReference, 
        bytes: usize, 
        allocator: AllocationSemantics
    );

    fn get_tls(&self) -> VMMutatorThread;

    fn barrier(&mut self) -> &mut dyn Barrier<VM>;

    unsafe fn barrier_impl<B: Barrier<VM>>(&mut self) -> &mut B;

    fn flush_remembered_sets(&mut self) { ... }

    fn flush(&mut self) { ... }
}

Each GC plan should provide their implementation of a MutatorContext. Note that this trait is no longer needed as we removed per-plan mutator implementation and we will remove this trait as well in the future.

Required methods

fn prepare(&mut self, tls: VMWorkerThread)

fn release(&mut self, tls: VMWorkerThread)

fn alloc(
&mut self,
size: usize,
align: usize,
offset: isize,
allocator: AllocationSemantics
) -> Address

fn post_alloc(
&mut self,
refer: ObjectReference,
bytes: usize,
allocator: AllocationSemantics
)

fn get_tls(&self) -> VMMutatorThread

fn barrier(&mut self) -> &mut dyn Barrier<VM>

Get active barrier trait object

unsafe fn barrier_impl<B: Barrier<VM>>(&mut self) -> &mut B

Force cast the barrier trait object to a concrete implementation.

Safety

The safety of this function is ensured by a down-cast check.

Provided methods

fn flush_remembered_sets(&mut self)

fn flush(&mut self)