Trait mmtk::scheduler::ProcessEdgesWork

pub trait ProcessEdgesWork: Send + 'static + Sized + DerefMut + Deref<Target = ProcessEdgesBase<Self::VM>> {
    type VM: VMBinding;
    type ScanObjectsWorkType: ScanObjectsWork<Self::VM>;

    const CAPACITY: usize;
    const OVERWRITE_REFERENCE: bool;
    const SCAN_OBJECTS_IMMEDIATELY: bool;

    fn new(
        edges: Vec<<<Self as ProcessEdgesWork>::VM as VMBinding>::VMEdge>, 
        roots: bool, 
        mmtk: &'static MMTK<Self::VM>
    ) -> Self;

    fn trace_object(&mut self, object: ObjectReference) -> ObjectReference;

    fn create_scan_work(
        &self, 
        nodes: Vec<ObjectReference>, 
        roots: bool
    ) -> Self::ScanObjectsWorkType;

    fn cache_roots_for_sanity_gc(&mut self) { ... }

    fn start_or_dispatch_scan_work(
        &mut self, 
        work_packet: impl GCWork<Self::VM>
    ) { ... }

    fn flush(&mut self) { ... }

    fn process_edge(
        &mut self, 
        slot: <<Self as ProcessEdgesWork>::VM as VMBinding>::VMEdge
    ) { ... }

    fn process_edges(&mut self) { ... }
}

Scan & update a list of object slots

Associated Types

type VM: VMBinding

type ScanObjectsWorkType: ScanObjectsWork<Self::VM>

The work packet type for scanning objects when using this ProcessEdgesWork.

Associated Constants

const CAPACITY: usize

const OVERWRITE_REFERENCE: bool

const SCAN_OBJECTS_IMMEDIATELY: bool

Required methods

fn new(
edges: Vec<<<Self as ProcessEdgesWork>::VM as VMBinding>::VMEdge>,
roots: bool,
mmtk: &'static MMTK<Self::VM>
) -> Self

fn trace_object(&mut self, object: ObjectReference) -> ObjectReference

Trace an MMTk object. The implementation should forward this call to the policy-specific trace_object() methods, depending on which space this object is in. If the object is not in any MMTk space, the implementation should forward the call to ActivePlan::vm_trace_object() to let the binding handle the tracing.

fn create_scan_work(
&self,
nodes: Vec<ObjectReference>,
roots: bool
) -> Self::ScanObjectsWorkType

Create an object-scanning work packet to be used for this ProcessEdgesWork.

roots indicates if we are creating a packet for root scanning. It is only true when this method is called to handle RootsWorkFactory::create_process_node_roots_work.

Provided methods

fn cache_roots_for_sanity_gc(&mut self)

fn start_or_dispatch_scan_work(&mut self, work_packet: impl GCWork<Self::VM>)

Start the a scan work packet. If SCAN_OBJECTS_IMMEDIATELY, the work packet will be executed immediately, in this method. Otherwise, the work packet will be added the Closure work bucket and will be dispatched later by the scheduler.

fn flush(&mut self)

Flush the nodes in ProcessEdgesBase, and create a ScanObjects work packet for it. If the node set is empty, this method will simply return with no work packet created.

fn process_edge(
&mut self,
slot: <<Self as ProcessEdgesWork>::VM as VMBinding>::VMEdge
)

fn process_edges(&mut self)