Trait mmtk::Plan

pub trait Plan: 'static + Sync + Downcast {
    type VM: VMBinding;

    fn constraints(&self) -> &'static PlanConstraints;

    fn base(&self) -> &BasePlan<Self::VM>;

    fn schedule_collection(
        &'static self, 
        _scheduler: &GCWorkScheduler<Self::VM>
    );

    fn get_spaces(&self) -> Vec<&dyn Space<Self::VM>>;

    fn get_allocator_mapping(
        &self
    ) -> &'static EnumMap<AllocationSemantics, AllocatorSelector>;

    fn prepare(&mut self, tls: VMWorkerThread);

    fn release(&mut self, tls: VMWorkerThread);

    fn collection_required(
        &self, 
        space_full: bool, 
        space: Option<&dyn Space<Self::VM>>
    ) -> bool;

    fn get_used_pages(&self) -> usize;

    fn create_copy_config(&'static self) -> CopyConfig<Self::VM> { ... }

    fn common(&self) -> &CommonPlan<Self::VM> { ... }

    fn generational(&self) -> &Gen<Self::VM> { ... }

    fn mmapper(&self) -> &'static FragmentedMapper { ... }

    fn options(&self) -> &Options { ... }

    fn is_current_gc_nursery(&self) -> bool { ... }

    fn enter_sanity(&self) { ... }

    fn leave_sanity(&self) { ... }

    fn is_in_sanity(&self) -> bool { ... }

    fn is_initialized(&self) -> bool { ... }

    fn should_trigger_gc_when_heap_is_full(&self) -> bool { ... }

    fn prepare_worker(&self, _worker: &mut GCWorker<Self::VM>) { ... }

    fn end_of_gc(&mut self, _tls: VMWorkerThread) { ... }

    fn get_reserved_pages(&self) -> usize { ... }

    fn get_total_pages(&self) -> usize { ... }

    fn get_available_pages(&self) -> usize { ... }

    fn get_mature_physical_pages_available(&self) -> usize { ... }

    fn get_collection_reserved_pages(&self) -> usize { ... }

    fn get_free_pages(&self) -> usize { ... }

    fn is_emergency_collection(&self) -> bool { ... }

    fn handle_user_collection_request(&self, tls: VMMutatorThread, force: bool) { ... }

    fn last_collection_was_exhaustive(&self) -> bool { ... }

    fn last_collection_full_heap(&self) -> bool { ... }

    fn force_full_heap_collection(&self) { ... }

    fn modify_check(&self, object: ObjectReference) { ... }

}

A plan describes the global core functionality for all memory management schemes. All global MMTk plans should implement this trait.

The global instance defines and manages static resources (such as memory and virtual memory resources).

Constructor:

For the constructor of a new plan, there are a few things the constructor must do (please check existing plans and see what they do in the constructor):

Create a HeapMeta, and use this HeapMeta to initialize all the spaces.

Create a vector of all the side metadata specs with SideMetadataContext::new_global_specs(), the parameter is a vector of global side metadata specs that are specific to the plan.

Initialize all the spaces the plan uses with the heap meta, and the global metadata specs vector.

Create a SideMetadataSanity object, and invoke verify_side_metadata_sanity() for each space (or invoke verify_side_metadata_sanity() in CommonPlan/BasePlan for the spaces in the common/base plan).

Methods in this trait:

Only methods that will be overridden by each specific plan should be included in this trait. The trait may provide a default implementation, and each plan can override the implementation. For methods that won’t be overridden, we should implement those methods in BasePlan (or CommonPlan) and call them from there instead. We should avoid having methods with the same name in both Plan and BasePlan, as this may confuse people, and they may call a wrong method by mistake.

Associated Types

type VM: VMBinding

Required methods

fn constraints(&self) -> &'static PlanConstraints

fn base(&self) -> &BasePlan<Self::VM>

fn schedule_collection(&'static self, _scheduler: &GCWorkScheduler<Self::VM>)

fn get_spaces(&self) -> Vec<&dyn Space<Self::VM>>

get all the spaces in the plan

fn get_allocator_mapping(
&self
) -> &'static EnumMap<AllocationSemantics, AllocatorSelector>

fn prepare(&mut self, tls: VMWorkerThread)

Prepare the plan before a GC. This is invoked in an initial step in the GC. This is invoked once per GC by one worker thread. tls is the worker thread that executes this method.

fn release(&mut self, tls: VMWorkerThread)

Release the plan after transitive closure. A plan can implement this method to call each policy’s release, or create any work packet that should be done in release. This is invoked once per GC by one worker thread. tls is the worker thread that executes this method.

fn collection_required(
&self,
space_full: bool,
space: Option<&dyn Space<Self::VM>>
) -> bool

Ask the plan if they would trigger a GC. If MMTk is in charge of triggering GCs, this method is called periodically during allocation. However, MMTk may delegate the GC triggering decision to the runtime, in which case, this method may not be called. This method returns true to trigger a collection.

Arguments

space_full: the allocation to a specific space failed, must recover pages within ‘space’.

space: an option to indicate if there is a space that has failed in an allocation.

fn get_used_pages(&self) -> usize

Get the number of pages that are used.

Provided methods

fn create_copy_config(&'static self) -> CopyConfig<Self::VM>

Create a copy config for this plan. A copying GC plan MUST override this method, and provide a valid config.

fn common(&self) -> &CommonPlan<Self::VM>

fn generational(&self) -> &Gen<Self::VM>

fn mmapper(&self) -> &'static FragmentedMapper

fn options(&self) -> &Options

fn is_current_gc_nursery(&self) -> bool

Is current GC only collecting objects allocated since last GC?

fn enter_sanity(&self)

fn leave_sanity(&self)

fn is_in_sanity(&self) -> bool

fn is_initialized(&self) -> bool

fn should_trigger_gc_when_heap_is_full(&self) -> bool

fn prepare_worker(&self, _worker: &mut GCWorker<Self::VM>)

Prepare a worker for a GC. Each worker has its own prepare method. This hook is for plan-specific per-worker preparation. This method is invoked once per worker by the worker thread passed as the argument.

fn end_of_gc(&mut self, _tls: VMWorkerThread)

Inform the plan about the end of a GC. It is guaranteed that there is no further work for this GC. This is invoked once per GC by one worker thread. tls is the worker thread that executes this method.

fn get_reserved_pages(&self) -> usize

Get the number of pages that are reserved, including used pages and pages that will be used (e.g. for copying).

fn get_total_pages(&self) -> usize

Get the total number of pages for the heap.

fn get_available_pages(&self) -> usize

Get the number of pages that are still available for use. The available pages should always be positive or 0.

fn get_mature_physical_pages_available(&self) -> usize

Return the number of pages available for allocation into the mature space. Only generational plans have to implement this function.

fn get_collection_reserved_pages(&self) -> usize

Get the number of pages that are reserved for collection. By default, we return 0. For copying plans, they need to override this and calculate required pages to complete a copying GC.

fn get_free_pages(&self) -> usize

Get the number of pages that are NOT used. This is clearly different from available pages. Free pages are unused, but some of them may have been reserved for some reason.

fn is_emergency_collection(&self) -> bool

fn handle_user_collection_request(&self, tls: VMMutatorThread, force: bool)

The application code has requested a collection.

fn last_collection_was_exhaustive(&self) -> bool

Return whether last GC was an exhaustive attempt to collect the heap. For many collectors this is the same as asking whether the last GC was a full heap collection.

fn last_collection_full_heap(&self) -> bool

Return whether last GC is a full GC.

fn force_full_heap_collection(&self)

Force the next collection to be full heap.

fn modify_check(&self, object: ObjectReference)

Implementations

impl<VM> dyn Plan<VM = VM> where
VM: Any + 'static,

pub fn is<__T: Plan<VM = VM>>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Plan<VM = VM>>(
self: Box<Self>
) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: Plan<VM = VM>>(
self: Rc<Self>
) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: Plan<VM = VM>>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: Plan<VM = VM>>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.