Module mmtk::util

Utilities used by other modules, including allocators, heap implementation, etc.

Re-exports

pub use self::address::Address;

pub use self::address::ObjectReference;

pub use self::reference_processor::ReferenceProcessor;

pub use self::opaque_pointer::*;

Modules

address

An abstract of memory address and object reference.

alloc

Allocators Various allocators implementation.

constants

Constants used in MMTk

conversions

Calculation, conversion and rounding for memory related numbers.

copy

The copy allocators for a GC worker.

is_mmtk_object

linear_scan

Linear scan through a heap range

malloc

Various malloc implementations (conditionally compiled by features)

memory

Wrapper functions for memory syscalls such as mmap, mprotect, etc.

metadata

Metadata (OnSide or InHeader) implementation. This is a generic module to work with metadata including side metadata and in-object metadata.

opaque_pointer

Opaque pointers used in MMTk, e.g. VMThread.

options

MMTk command line options.

reference_processor

Reference processing implementation.