Struct mmtk::util::address::ObjectReference

#[repr(transparent)]
pub struct ObjectReference(_);

ObjectReference represents address for an object. Compared with Address, operations allowed on ObjectReference are very limited. No address arithmetics are allowed for ObjectReference. The idea is from the paper High-level Low-level Programming (VEE09) and JikesRVM.

A runtime may define its “object references” differently. It may define an object reference as the address of an object, a handle that points to an indirection table entry where a pointer to the object is held, or anything else. Regardless, MMTk expects each object reference to have a pointer to the object (an address) in each object reference, and that address should be used for this ObjectReference type.

We currently do not allow an opaque ObjectReference type for which a binding can define their layout. We now only allow a binding to define their semantics through a set of methods in crate::vm::ObjectModel. Major refactoring is needed in MMTk to allow the opaque ObjectReference type, and we haven’t seen a use case for now.

Implementations

impl ObjectReference

pub const NULL: ObjectReference

pub fn to_raw_address(self) -> Address

Cast the object reference to its raw address. This method is mostly for the convinience of a binding.

MMTk should not make any assumption on the actual location of the address with the object reference. MMTk should not assume the address returned by this method is in our allocation. For the purposes of setting object metadata, MMTk should use crate::vm::ObjectModel::ref_to_address() or crate::vm::ObjectModel::ref_to_header().

pub fn from_raw_address(addr: Address) -> ObjectReference

Cast a raw address to an object reference. This method is mostly for the convinience of a binding. This is how a binding creates ObjectReference instances.

MMTk should not assume an arbitrary address can be turned into an object reference. MMTk can use crate::vm::ObjectModel::address_to_ref() to turn addresses that are from crate::vm::ObjectModel::ref_to_address() back to object.

pub fn to_address<VM: VMBinding>(self) -> Address

Get the in-heap address from an object reference. This method is used by MMTk to get an in-heap address for an object reference. This method is syntactic sugar for crate::vm::ObjectModel::ref_to_address. See the comments on crate::vm::ObjectModel::ref_to_address.

pub fn to_header<VM: VMBinding>(self) -> Address

Get the header base address from an object reference. This method is used by MMTk to get a base address for the object header, and access the object header. This method is syntactic sugar for crate::vm::ObjectModel::ref_to_header. See the comments on crate::vm::ObjectModel::ref_to_header.

pub fn to_object_start<VM: VMBinding>(self) -> Address

pub fn from_address<VM: VMBinding>(addr: Address) -> ObjectReference

Get the object reference from an address that is returned from crate::util::address::ObjectReference::to_address or crate::vm::ObjectModel::ref_to_address. This method is syntactic sugar for crate::vm::ObjectModel::address_to_ref. See the comments on crate::vm::ObjectModel::address_to_ref.

pub fn is_null(self) -> bool

is this object reference null reference?

pub fn value(self) -> usize

returns the ObjectReference

pub fn is_reachable(self) -> bool

Is the object reachable, determined by the policy? Note: Objects in ImmortalSpace may have is_live = true but are actually unreachable.

pub fn is_live(self) -> bool

Is the object live, determined by the policy?

pub fn is_movable(self) -> bool

pub fn get_forwarded_object(self) -> Option<Self>

Get forwarding pointer if the object is forwarded.

pub fn is_in_any_space(self) -> bool

pub fn is_sane(self) -> bool