Struct mmtk::util::metadata::header_metadata::HeaderMetadataSpec

pub struct HeaderMetadataSpec {
    pub bit_offset: isize,
    pub num_of_bits: usize,
}

This struct stores the specification of a header metadata bit-set. It supports either bits metadata of 1-7 bits in the same byte, or u8/u16/u32/u64 at an offset of their natural alignment.

For performance reasons, objects of this struct should be constants.

Fields
bit_offset: isizenum_of_bits: usize
Implementations

impl HeaderMetadataSpec

pub unsafe fn load<T: MetadataValue>(
&self,
header: Address,
optional_mask: Option<T>
) -> T

This function provides a default implementation for the load_metadata method from the ObjectModel trait.

Safety

This is a non-atomic load, thus not thread-safe.

pub fn load_atomic<T: MetadataValue>(
&self,
header: Address,
optional_mask: Option<T>,
ordering: Ordering
) -> T

This function provides a default implementation for the load_metadata_atomic method from the ObjectModel trait.

pub unsafe fn store<T: MetadataValue>(
&self,
header: Address,
val: T,
optional_mask: Option<T>
)

This function provides a default implementation for the store_metadata method from the ObjectModel trait.

Note: this function does compare-and-swap in a busy loop. So, unlike compare_exchange_metadata, this operation will always success.

Safety

This is a non-atomic store, thus not thread-safe.

pub fn store_atomic<T: MetadataValue>(
&self,
header: Address,
val: T,
optional_mask: Option<T>,
ordering: Ordering
)

This function provides a default implementation for the store_metadata_atomic method from the ObjectModel trait.

Note: this function does compare-and-swap in a busy loop. So, unlike compare_exchange_metadata, this operation will always success.

pub fn compare_exchange<T: MetadataValue>(
&self,
header: Address,
old_metadata: T,
new_metadata: T,
optional_mask: Option<T>,
success_order: Ordering,
failure_order: Ordering
) -> Result<T, T>

This function provides a default implementation for the compare_exchange_metadata method from the ObjectModel trait.

Note: this function only does fetch and exclusive store once, without any busy waiting in a loop.

pub fn fetch_add<T: MetadataValue>(
&self,
header: Address,
val: T,
order: Ordering
) -> T

This function provides a default implementation for the fetch_add method from the ObjectModel trait.

pub fn fetch_sub<T: MetadataValue>(
&self,
header: Address,
val: T,
order: Ordering
) -> T

This function provides a default implementation for the fetch_sub method from the ObjectModel trait.

pub fn fetch_and<T: MetadataValue>(
&self,
header: Address,
val: T,
order: Ordering
) -> T

This function provides a default implementation for the fetch_and method from the ObjectModel trait.

pub fn fetch_or<T: MetadataValue>(
&self,
header: Address,
val: T,
order: Ordering
) -> T

This function provides a default implementation for the fetch_or method from the ObjectModel trait.

pub fn fetch_update<T: MetadataValue, F: FnMut(T) -> Option<T> + Copy>(
&self,
header: Address,
set_order: Ordering,
fetch_order: Ordering,
f: F
) -> Result<T, T>

This function provides a default implementation for the fetch_update method from the ObjectModel trait. The semantics is the same as Rust’s fetch_update on atomic types.