Struct mmtk::util::metadata::side_metadata::SideMetadataSpec

pub struct SideMetadataSpec {
    pub name: &'static str,
    pub is_global: bool,
    pub offset: SideMetadataOffset,
    pub log_num_of_bits: usize,
    pub log_bytes_in_region: usize,
}

This struct stores the specification of a side metadata bit-set. It is used as an input to the (inline) functions provided by the side metadata module.

Each plan or policy which uses a metadata bit-set, needs to create an instance of this struct.

For performance reasons, objects of this struct should be constants.

Fields
name: &'static stris_global: booloffset: SideMetadataOffsetlog_num_of_bits: usize

Number of bits needed per region. E.g. 0 = 1 bit, 1 = 2 bit.

log_bytes_in_region: usize

Number of bytes of the region. E.g. 3 = 8 bytes, 12 = 4096 bytes (page).

Implementations

impl SideMetadataSpec

pub const fn is_absolute_offset(&self) -> bool

Is offset for this spec Address? (contiguous side metadata for 64 bits, and global specs in 32 bits)

pub const fn is_rel_offset(&self) -> bool

If offset for this spec relative? (chunked side metadata for local specs in 32 bits)

pub const fn get_absolute_offset(&self) -> Address

pub const fn get_rel_offset(&self) -> usize

pub const fn upper_bound_offset(&self) -> SideMetadataOffset

Return the upperbound offset for the side metadata. The next side metadata should be laid out at this offset.

pub const fn upper_bound_address_for_contiguous(&self) -> Address

The upper bound address for metadata address computed for this global spec. The computed metadata address should never be larger than this address. Otherwise, we are accessing the metadata that is laid out after this spec. This spec must be a contiguous side metadata spec (which uses address as offset).

pub fn bzero_metadata(&self, start: Address, size: usize)

Bulk-zero a specific metadata for a chunk. Note that this method is more sophisiticated than a simple memset, especially in the following cases:

the metadata for the range includes partial bytes (a few bits in the same byte).

for 32 bits local side metadata, the side metadata is stored in discontiguous chunks, we will have to bulk zero for each chunk’s side metadata.

Arguments

start: The starting address of a memory region. The side metadata starting from this data address will be zeroed.

size: The size of the memory region.

pub unsafe fn load<T: MetadataValue>(&self, data_addr: Address) -> T

Non-atomic load of metadata.

Safety

This is unsafe because:

Concurrent access to this operation is undefined behaviour.

Interleaving Non-atomic and atomic operations is undefined behaviour.

pub unsafe fn store<T: MetadataValue>(&self, data_addr: Address, metadata: T)

Non-atomic store of metadata.

Safety

This is unsafe because:

Concurrent access to this operation is undefined behaviour.

Interleaving Non-atomic and atomic operations is undefined behaviour.

pub fn load_atomic<T: MetadataValue>(
&self,
data_addr: Address,
order: Ordering
) -> T

pub fn store_atomic<T: MetadataValue>(
&self,
data_addr: Address,
metadata: T,
order: Ordering
)

pub fn compare_exchange_atomic<T: MetadataValue>(
&self,
data_addr: Address,
old_metadata: T,
new_metadata: T,
success_order: Ordering,
failure_order: Ordering
) -> Result<T, T>

pub fn fetch_add_atomic<T: MetadataValue>(
&self,
data_addr: Address,
val: T,
order: Ordering
) -> T

Wraps around on overflow.

pub fn fetch_sub_atomic<T: MetadataValue>(
&self,
data_addr: Address,
val: T,
order: Ordering
) -> T

pub fn fetch_and_atomic<T: MetadataValue>(
&self,
data_addr: Address,
val: T,
order: Ordering
) -> T

pub fn fetch_or_atomic<T: MetadataValue>(
&self,
data_addr: Address,
val: T,
order: Ordering
) -> T

pub fn fetch_update_atomic<T: MetadataValue, F: FnMut(T) -> Option<T> + Copy>(
&self,
data_addr: Address,
set_order: Ordering,
fetch_order: Ordering,
f: F
) -> Result<T, T>