Struct mmtk::util::reference_processor::ReferenceProcessors

pub struct ReferenceProcessors { /* private fields */ }

Holds all reference processors for each weak reference Semantics. Currently this is based on Java’s weak reference semantics (soft/weak/phantom). We should make changes to make this general rather than Java specific.

Implementations

impl ReferenceProcessors

pub fn new() -> Self

pub fn get(&self, semantics: Semantics) -> &ReferenceProcessor

pub fn add_soft_candidate<VM: VMBinding>(&self, reff: ObjectReference)

pub fn add_weak_candidate<VM: VMBinding>(&self, reff: ObjectReference)

pub fn add_phantom_candidate<VM: VMBinding>(&self, reff: ObjectReference)

pub fn enqueue_refs<VM: VMBinding>(&self, tls: VMWorkerThread)

This will invoke enqueue for each reference processor, which will call back to the VM to enqueue references whose referents are cleared in this GC.

pub fn forward_refs<E: ProcessEdgesWork>(
&self,
trace: &mut E,
mmtk: &'static MMTK<E::VM>
)

A separate reference forwarding step. Normally when we scan refs, we deal with forwarding. However, for some plans like mark compact, at the point we do ref scanning, we do not know the forwarding addresses yet, thus we cannot do forwarding during scan refs. And for those plans, this separate step is required.

pub fn scan_soft_refs<E: ProcessEdgesWork>(
&self,
trace: &mut E,
mmtk: &'static MMTK<E::VM>
)

Scan soft references.

pub fn scan_weak_refs<E: ProcessEdgesWork>(
&self,
trace: &mut E,
mmtk: &'static MMTK<E::VM>
)

Scan weak references.

pub fn scan_phantom_refs<E: ProcessEdgesWork>(
&self,
trace: &mut E,
mmtk: &'static MMTK<E::VM>
)

Scan phantom references.