Trait mmtk::vm::edge_shape::MemorySlice

pub trait MemorySlice: Send + Debug + PartialEq + Eq + Clone + Hash {
    type Edge: Edge;
    type EdgeIterator: Iterator<Item = Self::Edge>;
    fn iter_edges(&self) -> Self::EdgeIterator;

    fn start(&self) -> Address;

    fn bytes(&self) -> usize;

    fn copy(src: &Self, tgt: &Self);
}

A abstract memory slice represents a piece of heap memory.

Associated Types

type Edge: Edge

type EdgeIterator: Iterator<Item = Self::Edge>

Required methods

fn iter_edges(&self) -> Self::EdgeIterator

Iterate object edges within the slice. If there are non-reference values in the slice, the iterator should skip them.

fn start(&self) -> Address

Start address of the memory slice

fn bytes(&self) -> usize

Size of the memory slice

fn copy(src: &Self, tgt: &Self)

Memory copy support