Trait mmtk::vm::ObjectTracer

pub trait ObjectTracer {
    fn trace_object(&mut self, object: ObjectReference) -> ObjectReference;
}

Callback trait of scanning functions that directly trace through edges.

Required Methods

fn trace_object(&mut self, object: ObjectReference) -> ObjectReference

Call this function for the content of each edge, and assign the returned value back to the edge.

Note: This function is performance-critical. Implementations should consider inlining if necessary.