Trait mmtk::scheduler::GCWork

pub trait GCWork<VM: VMBinding>: 'static + Send {
    fn do_work(&mut self, worker: &mut GCWorker<VM>, mmtk: &'static MMTK<VM>);

    fn do_work_with_stat(
        &mut self, 
        worker: &mut GCWorker<VM>, 
        mmtk: &'static MMTK<VM>
    ) { ... }
}

Required methods

fn do_work(&mut self, worker: &mut GCWorker<VM>, mmtk: &'static MMTK<VM>)

Define the work for this packet. However, this is not supposed to be called directly. Usually do_work_with_stat() should be used.

Provided methods

fn do_work_with_stat(
&mut self,
worker: &mut GCWorker<VM>,
mmtk: &'static MMTK<VM>
)

Do work and collect statistics. This internally calls do_work(). In most cases, this should be called rather than do_work() so that MMTk can correctly collect statistics for the work packets. If the feature “work_packet_stats” is not enabled, this call simply forwards the call to do_work().