Trait mmtk::vm::Collection

pub trait Collection<VM: VMBinding> {
    fn stop_all_mutators<F>(tls: VMWorkerThread, mutator_visitor: F)
    where
        F: FnMut(&'static mut Mutator<VM>);
    fn resume_mutators(tls: VMWorkerThread);
    fn block_for_gc(tls: VMMutatorThread);
    fn spawn_gc_thread(tls: VMThread, ctx: GCThreadContext<VM>);
    fn prepare_mutator<T: MutatorContext<VM>>(
        tls_worker: VMWorkerThread,
        tls_mutator: VMMutatorThread,
        m: &T
    );

    fn out_of_memory(_tls: VMThread, err_kind: AllocationError) { ... }
    fn schedule_finalization(_tls: VMWorkerThread) { ... }
    fn post_forwarding(_tls: VMWorkerThread) { ... }
}

VM-specific methods for garbage collection.

Required Methods

fn stop_all_mutators<F>(tls: VMWorkerThread, mutator_visitor: F)where
F: FnMut(&'static mut Mutator<VM>),

Stop all the mutator threads. MMTk calls this method when it requires all the mutator to yield for a GC. This method is called by a single thread in MMTk (the GC controller). This method should not return until all the threads are yielded. The actual thread synchronization mechanism is up to the VM, and MMTk does not make assumptions on that.

Arguments:

tls: The thread pointer for the GC controller/coordinator.

fn resume_mutators(tls: VMWorkerThread)

Resume all the mutator threads, the opposite of the above. When a GC is finished, MMTk calls this method.

Arguments:

tls: The thread pointer for the GC controller/coordinator.

fn block_for_gc(tls: VMMutatorThread)

Block the current thread for GC. This is called when an allocation request cannot be fulfilled and a GC is needed. MMTk calls this method to inform the VM that the current thread needs to be blocked as a GC is going to happen. Then MMTk starts a GC. For a stop-the-world GC, MMTk will then call stop_all_mutators() before the GC, and call resume_mutators() after the GC.

Arguments:

tls: The current thread pointer that should be blocked. The VM can optionally check if the current thread matches tls.

fn spawn_gc_thread(tls: VMThread, ctx: GCThreadContext<VM>)

Ask the VM to spawn a GC thread for MMTk. A GC thread may later call into the VM through these VM traits. Some VMs have assumptions that those calls needs to be within VM internal threads. As a result, MMTk does not spawn GC threads itself to avoid breaking this kind of assumptions. MMTk calls this method to spawn GC threads during initialize_collection().

Arguments:

tls: The thread pointer for the parent thread that we spawn new threads from. This is the same tls when the VM calls initialize_collection() and passes as an argument.

ctx: The context for the GC thread.

If Controller is passed, it means spawning a thread to run as the GC controller. The spawned thread shall call memory_manager::start_control_collector.

If Worker is passed, it means spawning a thread to run as a GC worker. The spawned thread shall call memory_manager::start_worker. In either case, the Box inside should be passed back to the called function.

fn prepare_mutator<T: MutatorContext<VM>>(
tls_worker: VMWorkerThread,
tls_mutator: VMMutatorThread,
m: &T
)

Allow VM-specific behaviors for a mutator after all the mutators are stopped and before any actual GC work starts.

Arguments:

tls_worker: The thread pointer for the worker thread performing this call.

tls_mutator: The thread pointer for the target mutator thread.

m: The mutator context for the thread.

Provided Methods

fn out_of_memory(_tls: VMThread, err_kind: AllocationError)

Inform the VM of an out-of-memory error. The binding should hook into the VM’s error routine for OOM. Note that there are two different categories of OOM:

Critical OOM: This is the case where the OS is unable to mmap or acquire more memory. MMTk expects the VM to abort immediately if such an error is thrown.

Heap OOM: This is the case where the specified heap size is insufficient to execute the application. MMTk expects the binding to notify the VM about this OOM. MMTk makes no assumptions about whether the VM will continue executing or abort immediately.

See AllocationError for more information.

Arguments:

tls: The thread pointer for the mutator which failed the allocation and triggered the OOM.

err_kind: The type of OOM error that was encountered.

fn schedule_finalization(_tls: VMWorkerThread)

Inform the VM to schedule finalization threads.

Arguments:

tls: The thread pointer for the current GC thread.

fn post_forwarding(_tls: VMWorkerThread)

A hook for the VM to do work after forwarding objects.

This function is called after all of the following have finished:

The life and death of objects are determined. Objects determined to be live will not be reclaimed in this GC.

Live objects have been moved to their destinations. (copying GC only)

References in objects have been updated to point to new addresses. (copying GC only)

And this function may run concurrently with the release work of GC, i.e. freeing the space occupied by dead objects.

It is safe for the VM to read and write object fields at this time, although GC has not finished yet. GC will be reclaiming spaces of dead objects, but will not damage live objects. However, the VM cannot allocate new objects at this time.

One possible use of this hook is enqueuing {Soft,Weak,Phantom}Reference instances to reference queues (for Java). VMs (including JVM implementations) do not have to handle weak references this way, but mmtk-core provides this opportunity.

Arguments:

tls_worker: The thread pointer for the worker thread performing this call.